apps/ai-service/requirements.txt
# ─────────────────────────────────────────────────────
# Aegis Node — Python AI Microservice Dependencies
# Using flexible versions for Python 3.14 compatibility
# ─────────────────────────────────────────────────────

fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-genai>=1.5.0
pydantic>=2.11.0
python-dotenv>=1.0.0
httpx>=0.28.0
